refactor(YALB-639): namespaced organisms
