+++ components/03-organisms/layout/two-column/_two-column--example.twig
@@ -1,4 +1,4 @@
-{% extends "@organisms/layout/two-column/two-column.twig" %}
+{% extends "@organisms/layout/two-column/yds-two-column.twig" %}
   {% block two_column__primary %}
     {% include "@molecules/text/yds-text-field.twig" %}
   {% endblock %}
