+++ components/03-organisms/site-footer/_site-footer--examples.twig
@@ -1,6 +1,6 @@
 {% for theme, value in _context.themes %}
   <section>
-    {% include "./site-footer.twig" with {
+    {% include "./yds-site-footer.twig" with {
       site_footer__theme: theme,
     } %}
   </section><br />
