+++ components/03-organisms/site-header/_site-header--examples.twig
@@ -1,6 +1,6 @@
 {% for theme, value in _context.themes %}
   <section>
-    {% include "./site-header.twig" with {
+    {% include "./yds-site-header.twig" with {
       site_header__theme: theme,
     } %}
   </section><br />
