+++ components/03-organisms/site-header/yds-site-header.twig
@@ -37,7 +37,7 @@
     {# Primary Nav #}
     <div {{ bem('primary-nav', [], site_header__base_class) }}>
       {% block site_header__primary_nav %}
-        {% include "@organisms/menu/primary-nav/primary-nav.twig" %}
+        {% include "@organisms/menu/primary-nav/yds-primary-nav.twig" %}
       {% endblock %}
     </div>
   </div>
@@ -56,7 +56,7 @@
     {# Utility Nav #}
     <div {{ bem('utility-nav', [], site_header__base_class) }}>
       {% block site_header__utility_nav %}
-        {% include "@organisms/menu/utility-nav/utility-nav.twig" %}
+        {% include "@organisms/menu/utility-nav/yds-utility-nav.twig" %}
       {% endblock %}
     </div>
   </div>
