+++ components/05-page-examples/_intro-content-examples.twig
@@ -23,7 +23,7 @@
     text_with_image__width: 'highlight',
   } %}
 {% elseif intro_content == 'collection-featured' %}
-  {% include "@organisms/card-collection/card-collection.twig" with {
+  {% include "@organisms/card-collection/yds-card-collection.twig" with {
         card_example_type: 'news',
         card_collection__type: 'grid',
         card_collection__featured: 'true',
@@ -31,7 +31,7 @@
         card_collection__cards: [1, 2, 3],
   } %}
 {% elseif intro_content == 'collection-secondary' %}
-  {% include "@organisms/card-collection/card-collection.twig" with {
+  {% include "@organisms/card-collection/yds-card-collection.twig" with {
         card_example_type: 'news',
         card_collection__type: 'grid',
         card_collection__featured: 'false',
