+++ components/05-page-examples/miscellaneous/accordion-page.twig
@@ -1,8 +1,8 @@
 {% extends "@page-layouts/full-width.twig" %}
   {% block page__header %}
     {% include "@molecules/alert/yds-alert.twig" %}
-    {% include "@organisms/site-header/site-header.twig" %}
-    {% include "@organisms/menu/breadcrumbs/breadcrumbs.twig" %}
+    {% include "@organisms/site-header/yds-site-header.twig" %}
+    {% include "@organisms/menu/breadcrumbs/yds-breadcrumbs.twig" %}
   {% endblock %}
   {% block page__content %}
     {% include "@molecules/page-title//yds-page-title.twig" %}
