+++ components/05-page-examples/news/news-article.twig
@@ -1,7 +1,7 @@
 {% extends "@page-layouts/full-width.twig" %}
   {% block page__header %}
-    {% include "@organisms/site-header/site-header.twig" %}
-    {% include "@organisms/menu/breadcrumbs/breadcrumbs.twig" with {
+    {% include "@organisms/site-header/yds-site-header.twig" %}
+    {% include "@organisms/menu/breadcrumbs/yds-breadcrumbs.twig" with {
       breadcrumbs__items: [
         {title: 'Home', url: '#'},
         {title: 'Chem News', url: '#'},
